Move ReplicaCreateDropReason to transact.rs
