+++ src/adapter/src/catalog.rs
@@ -78,7 +78,9 @@ use uuid::Uuid;
 pub use crate::catalog::builtin_table_updates::BuiltinTableUpdate;
 pub use crate::catalog::open::BuiltinMigrationMetadata;
 pub use crate::catalog::state::CatalogState;
-pub use crate::catalog::transact::{DropObjectInfo, Op, TransactionResult};
+pub use crate::catalog::transact::{
+    DropObjectInfo, Op, ReplicaCreateDropReason, TransactionResult,
+};
 use crate::command::CatalogDump;
 use crate::coord::TargetCluster;
 use crate::session::{PreparedStatement, Session};
